コミットします。

+++ service/todo.go
@@ -3,6 +3,8 @@ package service
 import (
 	"context"
 	"database/sql"
+	"fmt"
+	"strings"
 
 	"github.com/TechBowl-japan/go-stations/model"
 )
@@ -166,7 +168,43 @@ func (s *TODOService) UpdateTODO(ctx context.Context, id int64, subject, descrip
 
 // DeleteTODO deletes TODOs on DB by ids.
 func (s *TODOService) DeleteTODO(ctx context.Context, ids []int64) error {
+	if len(ids) == 0 {
+		return nil
+	}
+
 	const deleteFmt = `DELETE FROM todos WHERE id IN (?%s)`
+
+	placeholders := strings.Repeat(",?", len(ids)-1)
+	delete := fmt.Sprintf(deleteFmt, placeholders)
+
+	stmt, err := s.db.PrepareContext(ctx, delete)
+	if err != nil {
+		return err
+	}
+	defer stmt.Close()
+
+	args := make([]interface{}, len(ids))
+	for i,id := range ids {
+		args[i] = id
+	}
+
+	result, err := stmt.ExecContext(ctx, args...)
+	if err != nil {
+		return err
+	}
+
+	rowsAffected, err := result.RowsAffected()
+	if err != nil {
+		return err
+	}
+
+	if rowsAffected == 0 {
+		return &model.ErrNotFound{}
+	}
+
+
+
+
 
 	return nil
 }
